% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtest-functions.R
\name{permute}
\alias{permute}
\title{Permute labels}
\usage{
permute(sampledata, grouping, sampletype)
}
\arguments{
\item{sampledata}{Data frame describing the samples.}

\item{grouping}{Grouping for repeated measures.}

\item{sampletype}{The sampletype used for testing (a column of sampledata).}
}
\value{
A permuted set of labels where the permutations are done
    over the levels of grouping.
}
\description{
Permutes sample labels, respecting repeated measures.
}
\keyword{internal}
