% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{phyloseqGraphTest-package}
\alias{phyloseqGraphTest-package}
\title{phyloseqGraphTest: Non-parametric graph-based testing for
microbiome data.}
\description{
This package lets you test for differences between groups of
samples with a graph-based permutation test.
}
\details{
The main function in the package is \code{\link{graph_perm_test}},
which takes a \code{\link[phyloseq]{phyloseq}} object.

The graph used in the test can be visualized using
\code{\link{plot_test_network}}. The permutation distribution and
the test statistic can be visualized with
\code{\link{plot_permutations}}.
}
