% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtest-functions.R
\name{plot_test_network}
\alias{plot_test_network}
\title{Plots the graph used for testing}
\usage{
plot_test_network(graphtest)
}
\arguments{
\item{graphtest}{The output from graph_perm_test.}
}
\value{
A ggplot object created by ggnetwork.
}
\description{
When using the graph_perm_test function, a graph is created. This
function will plot the graph used for testing with nodes colored by
sample type and edges marked as pure or mixed.
}

