% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtest-functions.R
\name{print.psgraphtest}
\alias{print.psgraphtest}
\title{Print psgraphtest objects}
\usage{
\method{print}{psgraphtest}(x, ...)
}
\arguments{
\item{x}{\code{psgraphtest} object.}

\item{...}{Not used}
}
\description{
Print psgraphtest objects
}
