% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtest-functions.R
\name{scale_safely}
\alias{scale_safely}
\title{Rescale x to (0, 1), except if x is constant}
\usage{
scale_safely(x, scale = diff(range(x)))
}
\arguments{
\item{x}{a vector to rescale}

\item{scale}{the scale on which to rescale the vector}
}
\value{
The rescaled vector, coerced to a vector if necessary.
  If the original vector was constant, all of its values are replaced by 0.5.
}
\description{
Copied from https://github.com/briatte/ggnetwork/blob/f3b8b84d28a65620a94f7aecd769c0ea939466e3/R/utilities.R so as to fix a problem with the cran version of ggnetwork.
}
\author{
Kipp Johnson
}
