% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtest-functions.R
\name{validGrouping}
\alias{validGrouping}
\title{Check for valid grouping}
\usage{
validGrouping(sd, sampletype, grouping)
}
\arguments{
\item{sd}{Data frame describing the samples.}

\item{sampletype}{The sampletype used for testing.}

\item{grouping}{Grouping for repeated measures.}
}
\value{
TRUE or FALSE for valid or invalid grouping.
}
\description{
Grouping should describe a repeated measures design, so this
function tests whether all of the levels of grouping have the same
value of sampletype.
}
\keyword{internal}
